dhra_api/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.2.0
certifi==2022.12.7
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-health-check==3.17.0
django-phonenumber-field==7.0.1
django-redis==5.2.0
django-rq==2.6.0
django-user-agents==0.4.0
djangorestframework==3.14.0
frozenlist==1.3.3
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
intelli-gateway==0.0.8
ipinfo==4.4.2
itypes==1.2.0
Jinja2==3.1.2
libmagic==1.0
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
multidict==6.0.3
numpy==1.24.2
phonenumbers==8.13.3
Pillow==9.3.0
psycopg2==2.9.5
pydantic==1.10.2
pydot==1.4.2
PyJWT==2.6.0
pyotp==2.8.0
pyparsing==3.0.9
python-decouple==3.6
python-dotenv==0.21.0
python-magic==0.4.27
pytz==2022.7
PyYAML==6.0
redis==4.4.0
requests==2.28.1
rq==1.11.1
sniffio==1.3.0
SQLAlchemy==1.4.45
sqlparse==0.4.3
starlette==0.22.0
typing_extensions==4.4.0
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.13
user-agents==2.2.0
uvloop==0.17.0
watchfiles==0.18.1
yarl==1.8.2
